+++ src/test/java/com/fasterxml/jackson/databind/deser/jdk/JDKScalarsTest.java
@@ -2,8 +2,6 @@ package com.fasterxml.jackson.databind.deser.jdk;
 
 import java.io.*;
 import java.lang.reflect.Array;
-import java.math.BigDecimal;
-import java.math.BigInteger;
 
 import org.junit.Assert;
 
@@ -130,35 +128,6 @@ public class JDKScalarsTest
         assertEquals(2, array.length);
         assertFalse(array[0]);
         assertFalse(array[1]);
-    }
-        
-    public void testBooleanPrimitiveArrayUnwrap() throws Exception
-    {
-        // [databind#381]
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.enable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        BooleanBean result = mapper.readValue(new StringReader("{\"v\":[true]}"), BooleanBean.class);
-        assertTrue(result._v);
-
-        try {
-            mapper.readValue("[{\"v\":[true,true]}]", BooleanBean.class);
-            fail("Did not throw exception while reading a value from a multi value array with UNWRAP_SINGLE_VALUE_ARRAY feature enabled");
-        } catch (JsonMappingException exp) {
-            verifyException(exp, "Attempted to unwrap");
-        }
-
-        result = mapper.readValue("{\"v\":[null]}", BooleanBean.class);
-        assertNotNull(result);
-        assertFalse(result._v);
-        
-        result = mapper.readValue("[{\"v\":[null]}]", BooleanBean.class);
-        assertNotNull(result);
-        assertFalse(result._v);
-        
-        boolean[] array = mapper.readValue(new StringReader("[ [ null ] ]"), boolean[].class);
-        assertNotNull(array);
-        assertEquals(1, array.length);
-        assertFalse(array[0]);
     }
 
     /**
@@ -602,100 +571,6 @@ public class JDKScalarsTest
                 quote("YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY3ODkwYWJjZGVmZ2hpamts\\nbW5vcHFyc3R1dnd4eXoxMjM0NTY3ODkwWA=="
         )));
     }    
-    /*
-    /**********************************************************
-    /* Simple non-primitive types
-    /**********************************************************
-     */
-
-    public void testSingleString() throws Exception
-    {
-        String value = "FOO!";
-        String result = MAPPER.readValue("\""+value+"\"", String.class);
-        assertEquals(value, result);
-    }
-    
-    public void testSingleStringWrapped() throws Exception
-    {
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.disable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        
-        String value = "FOO!";
-        try {
-            mapper.readValue("[\""+value+"\"]", String.class);
-            fail("Exception not thrown when attempting to unwrap a single value 'String' array into a simple String");
-        } catch (JsonMappingException exp) {
-            //exception thrown correctly
-        }
-        
-        mapper.enable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        
-        try {
-            mapper.readValue("[\""+value+"\",\""+value+"\"]", String.class);
-            fail("Exception not thrown when attempting to unwrap a single value 'String' array that contained more than one value into a simple String");
-        } catch (JsonMappingException exp) {
-            //exception thrown correctly
-        }
-        
-        String result = mapper.readValue("[\""+value+"\"]", String.class);
-        assertEquals(value, result);
-    }
-
-    public void testBigDecimal() throws Exception
-    {
-        final ObjectMapper mapper = objectMapper();
-        mapper.disable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        
-        BigDecimal value = new BigDecimal("0.001");
-        BigDecimal result = mapper.readValue(value.toString(), BigDecimal.class);
-        assertEquals(value, result);
-        try {
-            mapper.readValue("[" + value.toString() + "]", BigDecimal.class);
-            fail("Exception was not thrown when attempting to read a single value array of BigDecimal when UNWRAP_SINGLE_VALUE_ARRAYS feature is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        mapper.enable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        result = mapper.readValue("[" + value.toString() + "]", BigDecimal.class);
-        assertEquals(value, result);
-        
-        try {
-            mapper.readValue("[" + value.toString() + "," + value.toString() + "]", BigDecimal.class);
-            fail("Exception was not thrown when attempting to read a muti value array of BigDecimal when UNWRAP_SINGLE_VALUE_ARRAYS feature is enabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-    }
-
-    public void testBigInteger() throws Exception
-    {
-        final ObjectMapper mapper = objectMapper();
-        mapper.disable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        
-        BigInteger value = new BigInteger("-1234567890123456789012345567809");
-        BigInteger result = mapper.readValue(value.toString(), BigInteger.class);
-        assertEquals(value, result);
-        
-        //Issue#381
-        try {
-            mapper.readValue("[" + value.toString() + "]", BigInteger.class);
-            fail("Exception was not thrown when attempting to read a single value array of BigInteger when UNWRAP_SINGLE_VALUE_ARRAYS feature is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        mapper.enable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        result = mapper.readValue("[" + value.toString() + "]", BigInteger.class);
-        assertEquals(value, result);
-        
-        try {
-            mapper.readValue("[" + value.toString() + "," + value.toString() + "]", BigInteger.class);
-            fail("Exception was not thrown when attempting to read a muti value array of BigInteger when UNWRAP_SINGLE_VALUE_ARRAYS feature is enabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }        
-    }
 
     /*
     /**********************************************************
@@ -724,292 +599,6 @@ public class JDKScalarsTest
         jp.close();
     }
 
-    /*
-    /**********************************************************
-    /* Single-element as array tests
-    /**********************************************************
-     */
-    
-    // [databind#381]
-    public void testSingleElementScalarArrays() throws Exception {
-        final int intTest = 932832;
-        final double doubleTest = 32.3234;
-        final long longTest = 2374237428374293423L;
-        final short shortTest = (short) intTest;
-        final float floatTest = 84.3743f;
-        final byte byteTest = (byte) 43;
-        final char charTest = 'c';
-
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.enable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-
-        final int intValue = mapper.readValue(asArray(intTest), Integer.TYPE);
-        assertEquals(intTest, intValue);
-        final Integer integerWrapperValue = mapper.readValue(asArray(Integer.valueOf(intTest)), Integer.class);
-        assertEquals(Integer.valueOf(intTest), integerWrapperValue);
-
-        final double doubleValue = mapper.readValue(asArray(doubleTest), Double.class);
-        assertEquals(doubleTest, doubleValue);
-        final Double doubleWrapperValue = mapper.readValue(asArray(Double.valueOf(doubleTest)), Double.class);
-        assertEquals(Double.valueOf(doubleTest), doubleWrapperValue);
-
-        final long longValue = mapper.readValue(asArray(longTest), Long.TYPE);
-        assertEquals(longTest, longValue);
-        final Long longWrapperValue = mapper.readValue(asArray(Long.valueOf(longTest)), Long.class);
-        assertEquals(Long.valueOf(longTest), longWrapperValue);
-
-        final short shortValue = mapper.readValue(asArray(shortTest), Short.TYPE);
-        assertEquals(shortTest, shortValue);
-        final Short shortWrapperValue = mapper.readValue(asArray(Short.valueOf(shortTest)), Short.class);
-        assertEquals(Short.valueOf(shortTest), shortWrapperValue);
-
-        final float floatValue = mapper.readValue(asArray(floatTest), Float.TYPE);
-        assertEquals(floatTest, floatValue);
-        final Float floatWrapperValue = mapper.readValue(asArray(Float.valueOf(floatTest)), Float.class);
-        assertEquals(Float.valueOf(floatTest), floatWrapperValue);
-
-        final byte byteValue = mapper.readValue(asArray(byteTest), Byte.TYPE);
-        assertEquals(byteTest, byteValue);
-        final Byte byteWrapperValue = mapper.readValue(asArray(Byte.valueOf(byteTest)), Byte.class);
-        assertEquals(Byte.valueOf(byteTest), byteWrapperValue);
-
-        final char charValue = mapper.readValue(asArray(quote(String.valueOf(charTest))), Character.TYPE);
-        assertEquals(charTest, charValue);
-        final Character charWrapperValue = mapper.readValue(asArray(quote(String.valueOf(charTest))), Character.class);
-        assertEquals(Character.valueOf(charTest), charWrapperValue);
-
-        final boolean booleanTrueValue = mapper.readValue(asArray(true), Boolean.TYPE);
-        assertTrue(booleanTrueValue);
-
-        final boolean booleanFalseValue = mapper.readValue(asArray(false), Boolean.TYPE);
-        assertFalse(booleanFalseValue);
-
-        final Boolean booleanWrapperTrueValue = mapper.readValue(asArray(Boolean.valueOf(true)), Boolean.class);
-        assertEquals(Boolean.TRUE, booleanWrapperTrueValue);
-    }
-
-    public void testSingleElementArrayDisabled() throws Exception {
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.disable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        try {
-            mapper.readValue("[42]", Integer.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42]", Integer.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("[42.273]", Double.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42.2723]", Double.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("[42342342342342]", Long.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42342342342342342]", Long.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("[42]", Short.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42]", Short.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("[327.2323]", Float.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[82.81902]", Float.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("[22]", Byte.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[22]", Byte.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("['d']", Character.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("['d']", Character.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-
-        try {
-            mapper.readValue("[true]", Boolean.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[true]", Boolean.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-    }
-
-    public void testMultiValueArrayException() throws IOException {
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.enable(DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS);
-        
-        try {
-            mapper.readValue("[42,42]", Integer.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42,42]", Integer.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue("[42.273,42.273]", Double.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42.2723,42.273]", Double.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue("[42342342342342,42342342342342]", Long.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42342342342342342,42342342342342]", Long.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue("[42,42]", Short.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[42,42]", Short.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue("[327.2323,327.2323]", Float.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[82.81902,327.2323]", Float.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue("[22,23]", Byte.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[22,23]", Byte.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue(asArray(quote("c") + ","  + quote("d")), Character.class);
-            
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue(asArray(quote("c") + ","  + quote("d")), Character.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        
-        try {
-            mapper.readValue("[true,false]", Boolean.class);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-        try {
-            mapper.readValue("[true,false]", Boolean.TYPE);
-            fail("Single value array didn't throw an exception when DeserializationFeature.UNWRAP_SINGLE_VALUE_ARRAYS is disabled");
-        } catch (JsonMappingException exp) {
-            //Exception was thrown correctly
-        }
-    }
-
-    private static String asArray(Object value) {
-        final String stringVal = value.toString();
-        return new StringBuilder(stringVal.length() + 2).append("[").append(stringVal).append("]").toString();
-    }
 
     /*
     /**********************************************************
